Add a test to reproduce unidecode warning
